Improve calculation of labels
